+++ client-sdk/ts-web/rt/docs/changelog.md
@@ -6,6 +6,8 @@ New features:
 
 - `event.Visitor.visit` now returns whether it had a handler for the given
   event.
+- Added `transaction.SignatureMessageHandlersWithChainContext` for use with
+  `oasis.signature.visitMessage`.
 
 ## v0.2.0-alpha2